pyproject.toml
[tool.poetry]
name = "softmall-test-back"
version = "0.1.0"
description = ""
authors = ["Lee Constantine <lik42127@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
sqlalchemy = "^2.0.13"
asyncpg = "^0.27.0"
alembic = "^1.10.4"
python-dotenv = "^1.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
httpx = "^0.28.1"
pytest-dotenv = "^0.5.2"
faker = "^36.1.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.1"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.2.0"
flake8 = "^6.0.0"
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.0.0"
sqlalchemy = {extras = ["mypy"], version = "^2.0.12"}
types-python-jose = "^3.3.4.7"
types-passlib = "^1.7.7.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119

[tool.isort]
profile = "black"
line_length = 119
